pyproject.toml
[project]
name = "myownplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.16.5",
    "google-genai>=1.40.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "asyncpg",
    "gigachat",
    "greenlet",
]
